[refactor] POST FOLLOW errHandling 추가 Swagger에 반영

+++ src/following/service/following.service.ts
@@ -1,27 +1,42 @@
 import { Injectable } from '@nestjs/common';
 import { FollowFromTo } from '../../common/entities/FollowFromTo';
 import baseResponse from '../../common/utils/baseResponseStatus';
-import { sucResponse } from '../../common/utils/response';
+import { errResponse, sucResponse } from '../../common/utils/response';
+import { ProfilesRepository } from '../../profiles/profiles.repository';
 import { FollowingRepository } from '../following.repository';
 
 @Injectable()
 export class FollowingService {
     constructor(
         private followingRepository:FollowingRepository,
+        private profilesRepository:ProfilesRepository
     ){};
 
     async postFollow(fromProfileId: number,toProfileId: number): Promise<any> {//고쳐야함.
-
-        //팔로윙이 안되어있는 상태라면 팔로우 팔로우된 상태라면 언팔로우
-        const follow=new FollowFromTo(fromProfileId,toProfileId);
-        const isExist=await this.followingRepository.isExist(follow);
-        let foundEntity;
-        if(isExist.length==0){
-            foundEntity=await this.followingRepository.postFollow(follow);
-            return sucResponse(baseResponse.POST_FOLLOW);
-        }else{
-            await this.followingRepository.deleteFollow(follow);
-            return sucResponse(baseResponse.DELETE_FOLLOW);
+        try{
+            const isExistFromProfile=await this.profilesRepository.findProfileByProfileId(fromProfileId);
+            if(!isExistFromProfile){
+                console.log(isExistFromProfile);
+                return errResponse(baseResponse.FROM_PROFILE_ID_NOT_FOUND);
+            }
+            const isExistToProfile=await this.profilesRepository.findProfileByProfileId(toProfileId);
+            if(!isExistToProfile){
+                console.log(isExistToProfile);
+                return errResponse(baseResponse.TO_PROFILE_ID_NOT_FOUND);
+            }
+            //팔로윙이 안되어있는 상태라면 팔로우 팔로우된 상태라면 언팔로우
+            const follow=new FollowFromTo(fromProfileId,toProfileId);
+            const isExist=await this.followingRepository.isExist(follow);
+            let foundEntity;
+            if(isExist.length==0){
+                foundEntity=await this.followingRepository.postFollow(follow);
+                return sucResponse(baseResponse.POST_FOLLOW);
+            }else{
+                await this.followingRepository.deleteFollow(follow);
+                return sucResponse(baseResponse.DELETE_FOLLOW);
+            }
+        }catch(err){
+            return errResponse(baseResponse.DB_ERROR);
         }
     } 
 }
